requirements.txt
pygame~=2.5.2
torch~=2.1.2
numpy~=1.24.3
Pillow~=9.5.0